Update resource renaming again
