+++ de4dot.code/renamer/ResourceRenamer.cs
@@ -19,6 +19,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Text;
 using dnlib.DotNet;
 using dnlib.DotNet.Emit;
 using de4dot.blocks;
@@ -77,7 +78,7 @@ namespace de4dot.code.renamer {
 
 					if (!oldNameToTypeInfo.TryGetValue(codeString, out var typeInfo))
 						continue;
-					var newName = typeInfo.type.TypeDef.FullName;
+					var newName = EscapeTypeName(typeInfo.type.TypeDef.FullName);
 
 					bool renameCodeString = module.ObfuscatedFile.RenameResourcesInCode ||
 											IsCallingResourceManagerCtor(instrs, i, typeInfo);
@@ -143,18 +144,53 @@ namespace de4dot.code.renamer {
 		void RenameResources(List<TypeInfo> renamedTypes) {
 			var newNames = new Dictionary<Resource, RenameInfo>();
 			foreach (var info in renamedTypes) {
-				var oldFullName = info.oldFullName;
+				var oldFullName = EscapeTypeName(info.oldFullName);
 				if (!nameToResource.TryGetValue(oldFullName, out var resource))
 					continue;
 				if (newNames.ContainsKey(resource))
 					continue;
-				var newTypeName = info.type.TypeDef.FullName;
+				var newTypeName = EscapeTypeName(info.type.TypeDef.FullName);
 				var newName = newTypeName + resource.Name.String.Substring(oldFullName.Length);
 				newNames[resource] = new RenameInfo(resource, info, newName);
 
 				Logger.v("Renamed resource in resources: {0} => {1}", Utils.RemoveNewlines(resource.Name), newName);
 				resource.Name = newName;
 			}
+		}
+
+		static bool IsReservedTypeNameChar(char c) {
+			switch (c) {
+			case ',':
+			case '[':
+			case ']':
+			case '&':
+			case '*':
+			case '+':
+			case '\\':
+				return true;
+			default:
+				return false;
+			}
+		}
+
+		static bool HasReservedTypeNameChar(string s) {
+			foreach (var c in s) {
+				if (IsReservedTypeNameChar(c))
+					return true;
+			}
+			return false;
+		}
+
+		static string EscapeTypeName(string name) {
+			if (!HasReservedTypeNameChar(name))
+				return name;
+			var sb = new StringBuilder();
+			foreach (var c in name) {
+				if (IsReservedTypeNameChar(c))
+					sb.Append('\\');
+				sb.Append(c);
+			}
+			return sb.ToString();
 		}
 	}
 }
